refactor: backdgound color mapping 데이터 분리
