+++ src/utils/backgroundColor.js
@@ -1,0 +1,8 @@
+const colorMap = {
+  beige: 'var(--orange200)',
+  purple: 'var(--primary200)',
+  blue: 'var(--blue200)',
+  green: 'var(--green200)',
+};
+
+export default colorMap;
